requirements.txt
flask==1.1.2
flask_login==0.5.0
flask_migrate==2.7.0
flask_wtf==0.14.3
flask_sqlalchemy==2.4.4
sqlalchemy==1.3.23
email_validator==1.1.2
python-decouple==3.4
gunicorn==20.0.4
jinja2==2.11.3
pyjwt==2.1.0
